requirements.txt
django-modeltranslation==0.12.1
django>=2.0,<3.0
markdown>=2.6,<3.0
pymysql>=0.7,<1.0
django-webpack-loader>=0.5.0,<1.0
graphene-django>=2.0,<3.0
django-filter>=1.1,<2.0
channels>=1.1,<2.0
mysqlclient>=1.3,<2.0
asgi-redis>=1.4,<2.0
